Clean up unused styles and references
